requirements.txt
actionlib==1.11.13
angles==1.9.11
bondpy==1.7.20
cachetools==3.1.0
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
catkin==0.6.19
catkin-pkg==0.4.10
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
controller-manager==0.9.7
controller-manager-msgs==0.9.7
controller-manager-tests==0.9.7
cv-bridge==1.11.16
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
docutils==0.14
dynamic-reconfigure==1.5.50
gazebo-plugins==2.4.16
gazebo-ros==2.4.16
gencpp==0.5.5
genlisp==0.4.15
genmsg==0.5.8
genpy==0.5.10
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-speech==0.36.3
googleapis-common-protos==1.5.6
grpcio==1.18.0
idna==2.8
image-geometry==1.11.16
interactive-markers==1.11.4
laser-geometry==1.6.4
message-filters==1.11.21
numpy==1.16.1
pluginlib==1.10.7
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pyparsing==2.3.1
python-dateutil==2.8.0
python-qt-binding==0.2.19
pytz==2018.9
PyYAML==3.13
qt-dotgraph==0.2.33
qt-gui==0.2.33
qt-gui-cpp==0.2.33
qt-gui-py-common==0.2.33
requests==2.21.0
resource-retriever==1.11.8
rosbag==1.11.21
rosboost-cfg==1.11.14
rosclean==1.11.14
roscreate==1.11.14
rosgraph==1.11.21
roslaunch==1.11.21
roslib==1.11.14
roslint==0.10.0
roslz4==1.11.21
rosmake==1.11.14
rosmaster==1.11.21
rosmsg==1.11.21
rosnode==1.11.21
rosparam==1.11.21
rospkg==1.1.7
rospy==1.11.21
rosservice==1.11.21
rostest==1.11.21
rostopic==1.11.21
rosunit==1.11.14
roswtf==1.11.21
rqt-action==0.4.9
rqt-bag==0.4.8
rqt-bag-plugins==0.4.8
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.8
rqt-gui==0.4.8
rqt-gui-py==0.4.8
rqt-image-view==0.4.9
rqt-joint-trajectory-controller==0.9.4
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.4.8
rqt-py-console==0.4.8
rqt-reconfigure==0.4.10
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.7
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.5.8
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rsa==4.0
rviz==1.11.19
sensor-msgs==1.11.10
six==1.12.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.7.20
sounddevice==0.3.12
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.20
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
tf2-sensor-msgs==0.5.20
urllib3==1.24.1
xacro==1.9.5
